requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dj-database-url==0.5.0
Django==3.1.5
django-heroku==0.3.1
djangorestframework==3.12.2
gunicorn==20.0.4
heroku==0.1.4
idna==2.10
joblib==1.0.0
nltk==3.5
numpy==1.19.5
psycopg2==2.8.6
python-dateutil==1.5
pytz==2020.5
regex==2020.11.13
requests==2.25.1
sqlparse==0.4.1
torch==1.7.1
tqdm==4.56.0
typing-extensions==3.7.4.3
urllib3==1.26.3
whitenoise==5.2.0
